[feat] API filtering
This PR adds filtering of members for classes and components that allows
users to filter out private, protected, inherited, and deprecated
members. In the case of components, users can also filter out `internal`
properties - everything besides Yields and Arguments.

Filters only appear if there is something to filter, and by default
they are enabled (unchecked), so users should see the minimal public API
of the class.

The complexity of the filters is definitely not ideal, but at the moment
I'm unsure if pushing that complexity into the object model (e.g. by
adding `publicFields`, `privateFields`, `staticPublicFields`,
`staticPrivateFields`, etc directly to models) as part of the build step
will help much.