Remove application adapter and serializer
The presence of the application adapter and serializer broke ember-animated's test suite (because I was already using ember-data for another purpose).

In general I think it's risky for an addon to customize the global default application adapter & serializer, even when the consuming app is "just" an addon dummy app.

This PR limits the scope of the adapter and serializer customizations to only the models that are part of ember-cli-addon-docs.