fix array helper clobbering built in helper
The version of ember-composable-helpers currently being used provides an array helper.
ember@next has started erroring when it encounters built in helpers being clobbered.

ember-composable-helpers@3 has removed the array helper and upgraded to octane.
As an alternative to the commit, we could try to get a new version of ember-composable-helpers@2 cut that excludes the array helper for ember <= 3.8.