Use section instead of item for Introduction
This reflects how the addon-docs site itself is structured and is also consistent with the fact that the Quickstart guide later asks you to generate an index page, which would get added as a sub-item in the nav under the Introduction section. It's also convenient, because it subtly introduces the concept of a nav section, which isn't covered anywhere else in the Quickstart.