Fixes broken deployment with ember-auto-import@2
When using ember-auto-import@2, it will use the `rootURL` and use it to prefix the URLs of the injected chunks. It does this without adding a preceding slash, which means the URLs are invalid and chunks cannot be loaded. As a result, required dependencies, such as highlight.js, are not available (since they weren't loaded by the preceding chunk).